namespace Ocelot.AcceptanceTests.ServiceDiscovery
